mocktest_tasks/requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.9.0
